refactor: extract shared helpers to helpers.go, fix stdout consistency
Move jwtAuthPreRunE, getJSON, postJSON, getJSONWithHint, truncateUTF8,
and printList from course.go to dedicated helpers.go. Replace inline
JWT auth closure in project_task.go with shared function. Add project_id
stdout output to project owner update and register for composability.
